x/distribution; rename refresh to periodictask
